# docs/swagger.yaml
basePath: /api/v1
definitions:
  dtos.AlunoCreateDTO:
    properties:
      cpf:
        type: string
      nome:
        type: string
    required:
    - cpf
    - nome
    type: object
  dtos.AlunoUpdateDTO:
    properties:
      cpf:
        type: string
      nome:
        type: string
    type: object
  dtos.CursoCreateDTO:
    properties:
      faculdade_unique_id:
        type: integer
      nome:
        type: string
    required:
    - faculdade_unique_id
    - nome
    type: object
...
  dtos.CursoUpdateDTO:
    properties:
      faculdade_unique_id:
        type: integer
      nome:
        type: string
    type: object
...
  models.Aluno:
    properties:
      cpf:
        type: string
      disciplina_matriculas:
        items:
          $ref: '#/definitions/models.DisciplinaMatricula'
        type: array
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - cpf
    - nome
    type: object
  models.Curso:
    properties:
      curso_disciplinas:
        items:
          $ref: '#/definitions/models.CursoDisciplina'
        type: array
      faculdade:
        $ref: '#/definitions/models.Faculdade'
      faculdade_unique_id:
        type: integer
      nome:
        type: string
      unique_id:
        type: integer
    type: object
  models.CursoDisciplina:
    properties:
      curso:
        $ref: '#/definitions/models.Curso'
      curso_unique_id:
        type: integer
      disciplina:
        $ref: '#/definitions/models.Disciplina'
      disciplina_matriculas:
        items:
          $ref: '#/definitions/models.DisciplinaMatricula'
        type: array
      disciplina_unique_id:
        type: integer
      unique_id:
        type: integer
    type: object
  models.Disciplina:
    properties:
      curso_disciplinas:
        items:
          $ref: '#/definitions/models.CursoDisciplina'
        type: array
      nome:
        type: string
      professor:
        $ref: '#/definitions/models.Professor'
      professor_unique_id:
        type: integer
      unique_id:
        type: integer
    type: object
  models.DisciplinaMatricula:
    properties:
      aluno:
        $ref: '#/definitions/models.Aluno'
      aluno_unique_id:
        type: integer
      curso_disciplina:
        $ref: '#/definitions/models.CursoDisciplina'
      curso_disciplina_unique_id:
        type: integer
      data_matricula:
        type: string
      unique_id:
        type: integer
    type: object
  models.Faculdade:
    properties:
      cnpj:
        type: string
      cursos:
        items:
          $ref: '#/definitions/models.Curso'
        type: array
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - cnpj
    - nome
    type: object
  models.Professor:
    properties:
      disciplinas:
        items:
          $ref: '#/definitions/models.Disciplina'
        type: array
      formacao:
        type: string
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - formacao
    - nome
    type: object
host: localhost:5000
info:
  contact:
    name: Tathiane Souza
  description: Esta é a documentação API em Golang
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: API Faculdade passei direto
  version: "1.0"
paths:
  /aluno:
    get:
      consumes:
      - application/json
      description: Busca todos os alunos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
      summary: Busca todos os alunos
      tags:
      - Controller Aluno
    post:
      consumes:
      - application/json
      description: 'Cria um aluno, os campos que devem ser enviados são: nome e cpf'
      parameters:
      - description: aluno
        in: body
        name: aluno
        required: true
        schema:
          $ref: '#/definitions/dtos.AlunoCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Aluno'
      summary: Cria um aluno
      tags:
      - Controller Aluno
  /aluno/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um aluno
      parameters:
      - description: UniqueID do aluno
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta um aluno
      tags:
      - Controller Aluno
    get:
      consumes:
      - application/json
      description: Busca aluno por id
      parameters:
      - description: UniqueID do aluno
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
      summary: Busca aluno por id
      tags:
      - Controller Aluno
    put:
      consumes:
      - application/json
      description: 'Edita um aluno, os campos que podem ser enviados são: nome e cpf'
      parameters:
      - description: UniqueID do aluno
        in: path
        name: id
        required: true
        type: integer
      - description: aluno
        in: body
        name: aluno
        required: true
        schema:
          $ref: '#/definitions/dtos.AlunoUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
      summary: Edita um aluno
      tags:
      - Controller Aluno
  /aluno/curso/{id}:
    get:
      consumes:
      - application/json
      description: Realiza um filtro para obter todos os alunos de um curso
      parameters:
      - description: UniqueID do curso
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
      summary: Filtra alunos por curso
      tags:
      - Controller Aluno
  /aluno/disciplina/{id}:
    get:
      consumes:
      - application/json
      description: Realiza um filtro para obter todos os alunos de uma disciplina
      parameters:
      - description: UniqueID do disciplina
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
      summary: Filtra alunos por disciplina
      tags:
      - Controller Aluno
  /aluno/faculdade/{id}:
    get:
      consumes:
      - application/json
      description: Realiza um filtro para obter todos os alunos de uma faculdade
      parameters:
      - description: UniqueID da faculdade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
      summary: Filtra alunos por faculdade
      tags:
      - Controller Aluno
  /aluno/total/{id}:
    get:
      consumes:
      - application/json
      description: Realiza um filtro para obter o total de alunos de uma faculdade
      parameters:
      - description: UniqueID do faculdade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
      summary: Filtra o total de alunos de uma faculdade
      tags:
      - Controller Aluno
  /curso:
    get:
      consumes:
      - application/json
      description: Busca todos os cursos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Curso'
      summary: Buscar todos os cursos
      tags:
      - Controller Curso
    post:
      consumes:
      - application/json
      description: 'Cria um curso, os campos que devem ser enviados são: nome e "faculdade_unique_id'
      parameters:
      - description: curso
        in: body
        name: curso
        required: true
        schema:
          $ref: '#/definitions/dtos.CursoCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Curso'
      summary: Cria um curso
      tags:
      - Controller Curso
...
  /curso/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um curso
      parameters:
      - description: UniqueID do curso
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta um curso
      tags:
      - Controller Curso
    get:
      consumes:
      - application/json
      description: Busca curso por id
      parameters:
      - description: UniqueID do curso
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Curso'
      summary: Busca curso por id
      tags:
      - Controller Curso
    put:
      consumes:
      - application/json
      description: 'Edita um curso, os campos que podem ser enviados são: nome e "faculdade_unique_id'
      parameters:
      - description: UniqueID do curso
        in: path
        name: id
        required: true
        type: integer
      - description: curso
        in: body
        name: curso
        required: true
        schema:
          $ref: '#/definitions/dtos.CursoUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Curso'
      summary: Edita um curso
      tags:
      - Controller Curso
...
swagger: "2.0"